server/requirements.txt
Django==5.1.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.6.0
channels==4.1.0
uvicorn[standard]==0.30.6
psycopg[binary]==3.2.3
gunicorn==23.0.0
